refactor(element): move ChemicalElement and ElementSymbol types to dedicated types.ts
- Extract types from src/lib/index.ts to src/lib/element/types.ts
- Update imports in data.ts, labels.ts, state.svelte.ts
- Re-export types from element/index.ts for cleaner module structure
