+++ src/lib/element/data.ts
@@ -2,7 +2,7 @@
 // https://github.com/Bowserinator/Periodic-Table-JSON/blob/master/PeriodicTableJSON.json
 // https://gist.github.com/robertwb/22aa4dbfb6bcecd94f2176caa912b952
 
-import type { ChemicalElement } from '$lib'
+import type { ChemicalElement } from '$lib/element/types'
 
 export default [
   {
